Fixed line that generated particularly worrisome warning during compile on V100.

+++ src/mesh/mesh_refinement.cpp
@@ -250,7 +250,7 @@ void Mesh::ResetLoadBalanceCounters()
 {
   if (lb_automatic_) {
     for (int m=0; m<pmb_pack->nmb_thispack; ++m) {
-      costlist[pmb_pack->pmb->mb_gid.h_view(m)] = std::numeric_limits<double>::min();
+      costlist[pmb_pack->pmb->mb_gid.h_view(m)] = std::numeric_limits<float>::min();
     }
   }
   lb_flag_ = false;
